Added UI script and elements
For now the UI is drawn in the BoardControl script. Methods to change
elements of the UI will be in that script. The score field looks pretty
ghetto for until we update the GUI style. Also Board control is anchored
to am invisible Canvas object. Looked for solutions online and that is
what I found.